+++ wren-ai-service/src/pipelines/generation/followup_sql_generation.py
@@ -15,10 +15,11 @@ from src.core.provider import LLMProvider
 from src.pipelines.common import (
     TEXT_TO_SQL_RULES,
     SQLGenPostProcessor,
+    construct_instructions,
     sql_generation_system_prompt,
 )
 from src.utils import async_timer, timer
-from src.web.v1.services.ask import AskHistory
+from src.web.v1.services.ask import AskConfigurations, AskHistory
 
 logger = logging.getLogger("wren-ai-service")
 
@@ -116,6 +117,10 @@ Previous SQL Summary: {{ history.summary }}
 Previous SQL Query: {{ history.sql }}
 User's Follow-up Question: {{ query }}
 
+{% if instructions %}
+Instructions: {{ instructions }}
+{% endif %}
+
 Let's think step by step.
 """
 
@@ -129,12 +134,18 @@ def prompt(
     history: AskHistory,
     alert: str,
     prompt_builder: PromptBuilder,
+    configurations: AskConfigurations | None = None,
 ) -> dict:
     logger.debug(f"query: {query}")
     logger.debug(f"documents: {documents}")
     logger.debug(f"history: {history}")
+    logger.debug(f"configurations: {configurations}")
     return prompt_builder.run(
-        query=query, documents=documents, history=history, alert=alert
+        query=query,
+        documents=documents,
+        history=history,
+        alert=alert,
+        instructions=construct_instructions(configurations),
     )
 
 
@@ -193,6 +204,7 @@ class FollowUpSQLGeneration(BasicPipeline):
         contexts: List[str],
         history: AskHistory,
         project_id: str | None = None,
+        configurations: AskConfigurations | None = None,
     ) -> None:
         destination = "outputs/pipelines/generation"
         if not Path(destination).exists():
@@ -206,6 +218,7 @@ class FollowUpSQLGeneration(BasicPipeline):
                 "documents": contexts,
                 "history": history,
                 "project_id": project_id,
+                "configurations": configurations,
                 **self._components,
                 **self._configs,
             },
@@ -221,6 +234,7 @@ class FollowUpSQLGeneration(BasicPipeline):
         contexts: List[str],
         history: AskHistory,
         project_id: str | None = None,
+        configurations: AskConfigurations | None = None,
     ):
         logger.info("Follow-Up SQL Generation pipeline is running...")
         return await self._pipe.execute(
@@ -230,6 +244,7 @@ class FollowUpSQLGeneration(BasicPipeline):
                 "documents": contexts,
                 "history": history,
                 "project_id": project_id,
+                "configurations": configurations,
                 **self._components,
                 **self._configs,
             },
